# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mithra AI Family Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chats/{chatId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePair"
  /messages/{messageId}:
    patch:
      operationId: editMessage
      tags: [messages]
      summary: Edit a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageEdit"
      responses:
        "200":
          description: Edited message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /messages/{messageId}/regenerate:
    post:
      operationId: regenerateMessage
      tags: [messages]
      summary: Regenerate an AI response
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Regenerated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Files ────────────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List user's tasks
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user's notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, chatId, role, content, createdAt]
      properties:
        id:
          type: integer
        chatId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        model:
          type: ["string", "null"]
        tokensUsed:
          type: ["integer", "null"]
        createdAt:
          type: string
    MessagePair:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: "#/components/schemas/Message"
        assistantMessage:
          $ref: "#/components/schemas/Message"
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        model:
          type: ["string", "null"]
    MessageEdit:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    # ─── File ────────────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, userId, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        assignedToId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        isShared:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    TaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        isShared:
          type: boolean
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: ["string", "null"]
        assignedToId:
          type: ["integer", "null"]
        isShared:
          type: boolean

    # ─── Notification ────────────────────────────────────────────────────────
    Notification:
      type: object
      required: [id, userId, type, title, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fromUserId:
          type: ["integer", "null"]
        fromUserName:
          type: ["string", "null"]
        type:
          type: string
          enum: [message, task, file_share, system, family_update]
        title:
          type: string
        body:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string

    # ─── AI ──────────────────────────────────────────────────────────────────
...